Simplify structures
Summaries to own file for simpler class inheritance
