+++ marvel/creator.py
@@ -3,14 +3,13 @@
 __author__ = 'Garrett Pennington'
 __date__ = '02/07/14'
 
-from .core import MarvelObject, DataWrapper, DataContainer, Summary, List
-
+from .structures import DataWrapper, DataItem, DataContainer
 
 class CreatorDataWrapper(DataWrapper):
 
     @property
     def data(self):
-        return CreatorDataContainer(self.marvel, self.dict['data'])
+        return DataContainer(self.marvel, self.dict['data'], Creator)
 
     def next(self):
         return self._next(self.marvel.get_creators)
@@ -19,24 +18,13 @@ class CreatorDataWrapper(DataWrapper):
         return self._previous(self.marvel.get_creators)
 
 
-class CreatorDataContainer(DataContainer):
-
-    @property
-    def results(self):
-        return self.list_to_instance_list(self.dict['results'], Creator)
-
-
-class Creator(MarvelObject):
+class Creator(DataItem):
 
     """
     Creator object
     Takes a dict of creator attrs
     """
     _resource_url = 'creators'
-
-    @property
-    def id(self):
-        return int(self.dict['id'])
 
     @property
     def firstName(self):
@@ -57,18 +45,6 @@ class Creator(MarvelObject):
     @property
     def fullName(self):
         return self.dict['fullName']
-
-    @property
-    def modified(self):
-        return self.str_to_datetime(self.dict['modified'])
-
-    @property
-    def modified_raw(self):
-        return self.dict['modified']
-
-    @property
-    def resourceURI(self):
-        return self.dict['resourceURI']
 
     @property
     def urls(self):
@@ -95,61 +71,14 @@ class Creator(MarvelObject):
         return "%s.%s" % (self.dict['thumbnail']['path'], self.dict['thumbnail']['extension'])
 
     @property
-    def series(self):
-        """
-        Returns SeriesList object
-        """
-        from .series import SeriesList
-        return SeriesList(self.marvel, self.dict['series'])
+    def description(self):
+        raise AttributeError("'Creator' has no attribute description")
 
     @property
-    def stories(self):
-        """
-        Returns StoryList object
-        """
-        from .story import StoryList
-        return StoryList(self.marvel, self.dict['stories'])
-
-    @property
-    def comics(self):
-        from .comic import ComicList
-        """
-        Returns ComicList object
-        """
-        return ComicList(self.marvel, self.dict['comics'])
-
-    @property
-    def events(self):
-        """
-        Returns EventList object
-        """
-        from .event import EventList
-        return EventList(self.marvel, self.dict['events'])
+    def creators(self):
+        raise AttributeError("'Creator' has no attribute creators")
 
     def get_creators(self):
         raise AttributeError("'Creator' has no attribute get_creators")
 
 
-class CreatorList(List):
-
-    """
-    CreatorList object
-    """
-
-    @property
-    def items(self):
-        """
-        Returns List of CreatorSummary objects
-        """
-        return self.list_to_instance_list(self.dict['items'], CreatorSummary)
-
-
-class CreatorSummary(Summary):
-
-    """
-    CreatorSummary object
-    """
-
-    @property
-    def role(self):
-        return self.dict['role']
