+++ marvel/event.py
@@ -3,14 +3,15 @@
 __author__ = 'Garrett Pennington'
 __date__ = '02/07/14'
 
-from .core import MarvelObject, DataWrapper, DataContainer, Summary, List, Image
+from .structures import DataItem, DataWrapper, DataContainer, Image
+from .summaries import EventSummary
 
 
 class EventDataWrapper(DataWrapper):
 
     @property
     def data(self):
-        return EventDataContainer(self.marvel, self.dict['data'])
+        return DataContainer(self.marvel, self.dict['data'], Event)
 
     def next(self):
         return self._next(self.marvel.get_events)
@@ -19,14 +20,7 @@ class EventDataWrapper(DataWrapper):
         return self._previous(self.marvel.get_events)
 
 
-class EventDataContainer(DataContainer):
-
-    @property
-    def results(self):
-        return self.list_to_instance_list(self.dict['results'], Event)
-
-
-class Event(MarvelObject):
+class Event(DataItem):
 
     """
     Event object
@@ -35,23 +29,8 @@ class Event(MarvelObject):
     _resource_url = 'events'
 
     @property
-    def id(self):
-        return self.dict['id']
-
-    @property
     def title(self):
         return self.dict['title']
-
-    @property
-    def description(self):
-        """
-        :returns:  str -- The preferred description of the comic.
-        """
-        return self.dict['description']
-
-    @property
-    def resourceURI(self):
-        return self.dict['resourceURI']
 
     @property
     def urls(self):
@@ -86,34 +65,6 @@ class Event(MarvelObject):
         return Image(self.marvel, self.dict['thumbnail'])
 
     @property
-    def comics(self):
-        from .comic import ComicList
-        return ComicList(self.marvel, self.dict['comics'])
-
-    @property
-    def stories(self):
-        from .story import StoryList
-        return StoryList(self.marvel, self.dict['stories'])
-
-    @property
-    def series(self):
-        """
-        Returns SeriesList object
-        """
-        from .series import SeriesList
-        return SeriesList(self.marvel, self.dict['series'])
-
-    @property
-    def characters(self):
-        from .character import CharacterList
-        return CharacterList(self.marvel, self.dict['characters'])
-
-    @property
-    def creators(self):
-        from .creator import CreatorList
-        return CreatorList(self.marvel, self.dict['creators'])
-
-    @property
     def next(self):
         return EventSummary(self.marvel, self.dict['next'])
 
@@ -121,48 +72,9 @@ class Event(MarvelObject):
     def previous(self):
         return EventSummary(self.marvel, self.dict['previous'])
 
+    @property
+    def events(self):
+        raise AttributeError("'Event' has no attribute events")
+
     def get_events(self):
         raise AttributeError("'Event' has no attribute get_events")
-
-
-"""
-    Event {
-    id (int, optional): The unique ID of the event resource.,
-    title (string, optional): The title of the event.,
-    description (string, optional): A description of the event.,
-    resourceURI (string, optional): The canonical URL identifier for this resource.,
-    urls (Array[Url], optional): A set of public web site URLs for the event.,
-    modified (Date, optional): The date the resource was most recently modified.,
-    start (Date, optional): The date of publication of the first issue in this event.,
-    end (Date, optional): The date of publication of the last issue in this event.,
-    thumbnail (Image, optional): The representative image for this event.,
-    comics (ComicList, optional): A resource list containing the comics in this event.,
-    stories (StoryList, optional): A resource list containing the stories in this event.,
-    series (SeriesList, optional): A resource list containing the series in this event.,
-    characters (CharacterList, optional): A resource list containing the characters which appear in this event.,
-    creators (CreatorList, optional): A resource list containing creators whose work appears in this event.,
-    next (EventSummary, optional): A summary representation of the event which follows this event.,
-    previous (EventSummary, optional): A summary representation of the event which preceded this event.
-    }
-"""
-
-
-class EventList(List):
-
-    """
-    EventList object
-    """
-
-    @property
-    def items(self):
-        """
-        Returns List of EventSummary objects
-        """
-        return self.list_to_instance_list(self.dict['items'], EventSummary)
-
-
-class EventSummary(Summary):
-
-    """
-    EventSummary object
-    """
